requirements.txt
AuthEncoding==4.1
click==7.1.2
DateTime==4.3
Flask==0.12.2
Flask-HTTPAuth==4.0.0
flask-marshmallow==0.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.11.2
marshmallow==2.14.0
marshmallow-sqlalchemy==0.13.2
PyJWT==1.7.1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.16
uuid==1.30
Werkzeug==1.0.1
WSGIProxy2==0.4.6

